Separate settings, remove unnecessary code, fix logic
